+++ R/plot.R
@@ -19,7 +19,13 @@
 #' @param group A factor variable of the same number of
 #' rows as the data that is used to broduce grouped
 #' (faceted) plots of the posterior distribution.
-#' @param theme Any additional theme arguments to be added to ggplot2 (default is NULL).
+#' @param new_theme Any additional themes to be added to ggplot2 (default is NULL).
+#' @param outcome_label A character value that will replace the name of the outcome in the plot
+#' (default is the name of the response variable in the data frame).
+#' @param animate Whether to animate each posterior draw for continuous
+#' distributions (defaults to FALSE).
+#' @param reverse_bounds Whether to plot data using the original bounds in the data
+#' (i.e. not 0 and 1).
 #' @return If "both", prints both plots and returns a list of both plots as
 #' `ggplot2` objects. Otherwise, prints and returnst the specific plot
 #' as a `ggplot2` object.
@@ -47,12 +53,16 @@
 #' @import ggplot2
 #' @importFrom dplyr select
 #' @importFrom utils packageVersion
+#' @importFrom gganimate transition_time ease_aes
 pp_check_ordbeta <- function(model=NULL,
                              type="both",
                              ndraws=10,
                              cores=NULL,
                              group=NULL,
-                             theme=NULL) {
+                             new_theme=NULL,
+                             outcome_label=NULL,
+                             animate=FALSE,
+                             reverse_bounds=TRUE) {
 
     #define globals
 
@@ -63,6 +73,47 @@ pp_check_ordbeta <- function(model=NULL,
     full_dist <- posterior_predict(model,ndraws=ndraws,cores=NULL)
 
     outcome <- model$data[[1]]
+
+    if(!is.null(outcome_label)) {
+
+      outcome_label <- outcome_label
+
+    } else {
+
+      outcome_label <- names(model$data)[1]
+
+    }
+
+    if(reverse_bounds) {
+
+        # revert to original scale
+
+        l_bound <- model$lower_bound
+        up_bound <- model$upper_bound
+
+
+        full_dist <- apply(full_dist, 1, function(c) {
+
+          c <- (c *(up_bound - l_bound)) + l_bound
+
+          c
+
+        }) %>% t
+
+        outcome <- (outcome *(up_bound - l_bound)) + l_bound
+        # deal with floating point errors
+
+        tol <- 1e-5
+
+        outcome[abs(outcome - up_bound) < tol] <- up_bound
+        outcome[abs(outcome - l_bound) < tol] <- l_bound
+
+    } else {
+
+      l_bound <- 0
+      up_bound <- 1
+
+    }
 
     output <- list()
 
@@ -79,27 +130,56 @@ pp_check_ordbeta <- function(model=NULL,
 
                 apply(full_dist[,group==g], 1, function(c) {
 
-                  tibble(num_ones=sum(c==1),
-                         num_zeroes=sum(c==0),
+                  tibble(num_ones=sum(c==up_bound),
+                         num_zeroes=sum(c==l_bound),
+                         num_cont=sum(c<up_bound & c>l_bound),
                          group=g)
 
                 }) %>% bind_rows
 
             }) %>% bind_rows
 
-          plot_data_bar <- bind_rows(select(plot_data_bar, var="num_ones", group) %>% mutate(true=sum(outcome==1,na.rm=T),type=1),
-                                     select(plot_data_bar, var="num_zeroes", group) %>% mutate(true=sum(outcome==0,na.rm=T),type=0))
-          true_data_bar <- tibble(type=c(0,1),
-                                  true=c(sum(outcome==0,na.rm=T),sum(outcome==1,na.rm=T)))
+          plot_data_bar$iter <- 1:nrow(plot_data_bar)
+
+          plot_data_bar <- bind_rows(select(plot_data_bar, var="num_ones", group) %>% mutate(true=sum(outcome==up_bound,na.rm=T),
+                                                                                             type=as.character(round(up_bound,3))),
+                                     select(plot_data_bar, var="num_zeroes", group) %>% mutate(true=sum(outcome==l_bound,na.rm=T),
+                                                                                               type=as.character(round(l_bound,3))),
+                                     select(plot_data_bar, var="num_cont", group) %>% mutate(true=sum(outcome>l_bound & outcome<up_bound,na.rm=T),
+                                                                                                   type=paste0('(',round(l_bound,3),
+                                                                                                               ',',round(up_bound,3),
+                                                                                                               ')'))) %>%
+                          mutate(type=factor(type,
+                                             levels=c(round(l_bound,3),
+                                                      paste0('(',round(l_bound,3),
+                                                             ',',round(up_bound,3),
+                                                             ')'),
+                                 round(up_bound,3))))
+
+          true_data_bar <- tibble(type=c(round(l_bound,3),
+                                         round(up_bound,3),
+                                         paste0('(',round(l_bound,3),
+                                                                 ',',round(up_bound,3),
+                                                                 ')')),
+                                  true=c(sum(outcome==l_bound,na.rm=T),
+                                         sum(outcome==up_bound,na.rm=T),
+                                         sum(outcome>l_bound & outcome<up_bound,na.rm=T))) %>%
+                                  mutate(type=factor(type,
+                                                     levels=c(round(l_bound,3),
+                                                              paste0('(',round(l_bound,3),
+                                                                     ',',round(up_bound,3),
+                                                                     ')'),
+                                                              round(up_bound,3))))
+
           bar_plot <- ggplot(plot_data_bar,
                              aes(x=type,y=var)) +
             geom_col(data=true_data_bar, aes(y=true),width=.5,fill="gray") +
             stat_summary(fun.data="median_hilow",geom = "pointrange") +
             labs(y="Observed and Estimated Counts",
-                 x=paste0("Lower/Higher Discrete Bounds in ", names(model$data)[1]),
+                 x=paste0("No. of Discrete and Continuous Observations in ", outcome_label),
                  caption="Estimates are posterior medians with 5% and 95% high/low quantiles.\nCount of discrete values in data are represented by gray bars.") +
-            ggtitle("Posterior Predictions for Discrete Values",
-                    subtitle=paste0("Outcome: ",names(model$data)[1])) +
+            ggtitle("Posterior Predictions for Types of Responses in Data",
+                    subtitle=paste0("Outcome: ",outcome_label)) +
             theme_minimal()
 
           # add in facets if groups>1
@@ -110,7 +190,7 @@ pp_check_ordbeta <- function(model=NULL,
           }
 
 
-          output$discrete <- bar_plot
+          output$discrete <- bar_plot + new_theme
 
     }
 
@@ -119,6 +199,29 @@ pp_check_ordbeta <- function(model=NULL,
       # need to change the posterior predict function to get the argument from ...
 
       cont_dist <- posterior_predict(model,ndraws=ndraws,cores=NULL,ntrys=100)
+
+      if(reverse_bounds) {
+
+        # revert to original scale
+
+        l_bound <- model$lower_bound
+        up_bound <- model$upper_bound
+
+
+        cont_dist <- apply(cont_dist, 1, function(c) {
+
+          c <- (c *(up_bound - l_bound)) + l_bound
+
+          c
+
+        }) %>% t
+
+      } else {
+
+        l_bound <- 0
+        up_bound <- 1
+
+      }
 
       # if prediction was continuous, get continuous value from the outcome
 
@@ -130,8 +233,7 @@ pp_check_ordbeta <- function(model=NULL,
 
           counter <<- counter + 1
 
-          tibble(pred=col[!(outcome %in% c(0,1))],
-                          true=outcome[!(outcome %in% c(0,1))],
+          tibble(pred=col[!(col %in% c(l_bound,up_bound))],
                           draw=counter,
                  group=g)
 
@@ -139,27 +241,43 @@ pp_check_ordbeta <- function(model=NULL,
 
       }) %>% bind_rows
 
+      true_data <- tibble(true=outcome[!(outcome %in% c(l_bound,up_bound))])
 
-      cont_plot <- plot_data_dens %>%
-        ggplot(aes(x=pred,group=draw)) +
-        geom_density(alpha=0.7) +
-        theme_minimal() +
-        labs(y="Probability Density",x=paste0("Continuous values of ",names(model$data)[1]),
-             caption="Each black line is a draw from the posterior distribution.\nData distribution represented by gray line.") +
-        ggtitle("Posterior Predictions for Continuous Data",
-                subtitle=paste0("Outcome: ",names(model$data)[1]))
+      if(animate) {
 
-      # use different syntax depending on version of ggplot2
 
-      if(packageVersion('ggplot2')=="3.4.0") {
-
-        cont_plot <- cont_plot + geom_density(aes(x=true),linewidth=2,colour="gray",alpha=0.7)
+        cont_plot <- plot_data_dens %>%
+          ggplot(aes(x=pred)) +
+          geom_density(alpha=0.7,bounds=c(l_bound,up_bound)) +
+          theme_minimal() +
+          labs(y="Probability Density",x=paste0("Continuous values of ",outcome_label),
+               caption="Each black line is a draw from the posterior distribution.\nData distribution represented by gray line.") +
+          transition_time(draw) +
+          ease_aes('elastic-in-out') +
+          ggtitle(paste0("Posterior Predictions for Data Bounded from ",
+                         round(l_bound,3), " to ", round(up_bound,3)),
+                  subtitle=paste0("Outcome: ",outcome_label))
 
       } else {
 
-        cont_plot <- cont_plot + geom_density(aes(x=true),size=2,colour="gray",alpha=0.7)
+        cont_plot <- plot_data_dens %>%
+          ggplot(aes(x=pred,group=draw)) +
+          geom_density(alpha=0.7,bounds=c(l_bound,up_bound)) +
+          theme_minimal() +
+          labs(y="Probability Density",x=paste0("Continuous values of ",outcome_label),
+               caption="Each black line is a draw from the posterior distribution.\nData distribution represented by gray line.") +
+          ggtitle(paste0("Posterior Predictions for Data Bounded from ",
+                         round(l_bound,3), " to ", round(up_bound,3)),
+                  subtitle=paste0("Outcome: ",outcome_label))
+
 
       }
+
+      cont_plot <- cont_plot + geom_density(data=true_data,
+                                            mapping=aes(x=true),
+                                            linewidth=2,colour="gray",alpha=0.7,
+                                            bounds=c(l_bound,up_bound)) +
+        new_theme
 
       output$continuous <- cont_plot